moved types from weight.ts to types file
