--- apps/gym/src/app/(user)/weight/Weight.ts
@@ -1,8 +1,0 @@
-export interface EntryModalProps {
-  closeModal: () => void
-  deleteEntry: () => void
-}
-
-export interface AddWeightModalProps {
-  closeModal: () => void
-}
